buffer: separate PeriodBufferManager and PeriodBuffer
